requirements.txt
PyMuPDF==1.22.5
pyttsx3==2.90
Pillow==10.0.0
pygame==2.5.0